lettuce: overlay: Set the minimum brightness to 2
Previous to P, this was inferred as the minimum brightness because the
dark configuration was unset (and defaulted to 1). With the removal of
the dark configuration, the minimum brightness is actually respected,
but is now higher than desired. Setting this to 1 returns us to the
state we were in pre-P.

[ghostrider-reborn: 1 is too dark on lettuce so let's use 2]

Change-Id: I815e928c3294ceda1d2d0d11082ce75345b5e2e3
Fixes: 79660591
Test: manual